I just converted the budget tracker to view/set endpoints

+++ budget_tracker.py
@@ -1,4 +1,7 @@
 import json
+import os.path
+
+
 class BudgetTracker:
     def __init__(self):
         self.feeding_limit = int(input("Enter your monthly feeding budget "))
@@ -11,19 +14,34 @@ class BudgetTracker:
 
 
     def save_budget_data(self):
+        new_budget(budget=self.budget, feeding=self.feeding_limit, airtime_data=self.airtime_data_limit, electricity=self.electricity_limit,
+                   betting=self.betting_limit, transfer=self.transfer_limit)
+
+
+def new_budget(budget, feeding, airtime_data, electricity, betting, transfer):
+    if os.path.exists("budget_limits.txt"):
+        return {"Message" : "You have budget limits already, view or edit them"}
+    else:
         budget_limit_dict = {
-            "budget": self.budget,
-            "feeding": self.feeding_limit,
-            "airtime/data": self.airtime_data_limit,
-            "electricity": self.electricity_limit,
-            "betting": self.betting_limit,
-            "transfer": self.transfer_limit,
+            "budget": budget,
+            "feeding": feeding,
+            "airtime/data": airtime_data,
+            "electricity": electricity,
+            "betting": betting,
+            "transfer": transfer,
         }
-        # save budget data
+
         with open("budget_limits.txt", "w") as data:
             json.dump(budget_limit_dict, data)
+        return {"message": "Budgets set successfully!"}
 
-
+def get_budget():
+    try:
+        with open("budget_limits.txt", "r") as budget_data:
+            budget_file = json.load(budget_data)
+            return budget_file
+    except FileNotFoundError:
+        return {"message": "You have not set your budget limits yet"}
 
 def check_category_limits(limit, category_amount, budget_dict):
     if budget_dict[limit] <= category_amount:
